refactor: change float to double

+++ internal/token/token.go
@@ -8,7 +8,7 @@ const (
 
 	STRING = "STRING"
 	INT    = "INT"
-	FLOAT  = "FLOAT"
+	DOUBLE = "DOUBLE"
 	IDENT  = "IDENT"
 
 	LBRACE   = "{"
